Update recipe for just pyside2, not shiboken

+++ recipe/build.sh
@@ -7,24 +7,6 @@ then
   XVFB_RUN="xvfb-run -s '-screen 0 640x480x24'"
 fi
 
-pushd sources/shiboken2
-mkdir build && cd build
-
-cmake \
-  -DCMAKE_PREFIX_PATH=${PREFIX} \
-  -DCMAKE_INSTALL_PREFIX=${PREFIX} \
-  -DCMAKE_BUILD_TYPE=Release \
-  -DCMAKE_INSTALL_RPATH=${PREFIX}/lib \
-  -DBUILD_TESTS=OFF \
-  -DPYTHON_EXECUTABLE=${PYTHON} \
-  ..
-make install -j${CPU_COUNT}
-popd
-
-${PYTHON} setup.py dist_info --build-type=shiboken2
-_pythonpath=`${PYTHON} -c "from sysconfig import get_python_version; print(get_python_version())"`
-cp -r shiboken2-5.15.3.dist-info "${PREFIX}"/lib/python"${_pythonpath}"/site-packages/
-
 pushd sources/pyside2
 mkdir build && cd build
 
@@ -64,5 +46,8 @@ mv ${PREFIX}/bin/pyside_tool.py "${PREFIX}"/lib/python"${_pythonpath}"/site-pack
 # This is just an extra file that doesn't do anything. Shiboken Doesn't have any entry points
 rm ${PREFIX}/bin/shiboken_tool.py
 
-rm -rf ${PREFIX}/include/qt/xcb
+if test `uname` = "Linux"
+then
+  rm -rf ${PREFIX}/include/qt/xcb
+fi
